+++ CHANGELOG.md
@@ -5,9 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2025-07-31
+
+### Fixed
+
+- Fixed Rinegine path checking on Windows
+
+## [0.1.6] - 2025-07-14
+
+### Changed
+
+- Changed path for executable file for Linux.
+
+## [0.1.5] - 2025-07-07
+
+### Fixed
+
+- Fixed paths to work properly on Linux
+
 ## [0.1.4] - 2025-05-14
 
-###
+### Fixed
 
 - Fixed a bug with paths. The path to Rinegine was accepted only with \